simplify sign in message regex, refactor internal files
