+++ packages/core/util/src/index.ts
@@ -1,3 +1,4 @@
 export * from './commitment.js';
 export * from './endpoint.js';
-export * from './verify.js';
+export * from './signIn.js';
+export * from './signMessage.js';
